install chef via ruby, add option for rvm (disabled)
This is just a quick and dirty change to get ubuntu-12.04 built with
chef install via the system ruby. Omnibus is a no-go for the postgresql
cookbook. *mega sad face*

rvm added, but commented out. It was originally intended for use with
test-kitchen, but you're gunna have a bad time if you're using the rvm
cookbook in your tests. Patching test-kitchen to use system ruby is a
better alternative